Testo de landingpage y contacto

+++ database/migrations/2022_09_12_212654_create_contactos_table.php
@@ -17,7 +17,7 @@ class CreateContactosTable extends Migration
             $table->id();
             $table->string('name', 100);
             $table->string('email');
-            $table->integer('telefono');
+            $table->bigInteger('telefono');
             $table->text('message', 255);
             $table->timestamps();
         });
